create a lookup table from drug concepts in vocabs RxNorm and/or RxNorm Extension to ATC drug classes

EXPERIMENTAL either all drug concepts filtered by concept_class_id OR all drug concepts in a passed table (e.g. drug_exposure) ATC is a WHO drug classification incorporated within RxNorm. Anatomical Therapeutic Chemical Classification System.

Usage

omop_drug_lookup_create(
  df = NULL,
  name_drug_concept_id = "drug_concept_id",
  concept_class_ids = c("Ingredient"),
  drug_concept_vocabs = c("RxNorm", "RxNorm Extension"),
  outfile = NULL,
  messages = TRUE
)

Arguments

df

optional table containing drug concept ids

name_drug_concept_id

optional name of column containing drug concept ids, default="drug_concept_id"

concept_class_ids

optional filter of concept_class_ids, multiple allowed, default = "Ingredient", ignored if a table is passed as df

drug_concept_vocabs

vocabs containing drug concepts default = c("RxNorm","RxNorm Extension") option just = "RxNorm Extension"

outfile

name for output file default=NULL for no file

messages

whether to print info messages, default=TRUE

Value

data frame with drug concepts and ATC classes

Examples

#to create a lookup table for all Ingredients
drug_lookup = omop_drug_lookup_create()
#> creating drug lookup may take more than a few seconds (e.g. ~20s for all concept_class_ids == 'Ingredient'
#counting numbers of concepts under each level in ATC hierarchy