find immediate omop concept relations of one passed ADVISORY instead to use omop_relations() with default nsteps=1 arg.

find immediate omop concept relations of one passed ADVISORY instead to use omop_relations() with default nsteps=1 arg.

Usage

omop_relations1step(
  c_id = NULL,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  r_ids = NULL,
  itself = FALSE,
  names2avoid = c("SNOMED CT core", "Defined", "Primitive"),
  join_names = TRUE,
  messages = TRUE
)

Arguments

c_id

single omop concept_id or exact concept_name to get relations of, default NULL returns all

c_ids

one or more concept_id to filter by, default NULL for all

d_ids

one or more domain_id to filter by, default NULL for all

v_ids

one or more vocabulary_id to filter by, default NULL for all

cc_ids

one or more concept_class_id to filter by, default NULL for all

standard

one or more standard_concept to filter by, default NULL for all, S,C

r_ids

one or more relationship_id to filter by, default NULL for all, e.g c('Is a','Subsumes')

itself

whether to include relations to concept itself, default=FALSE

names2avoid

concept names to avoid, defaults to generic concepts with lots relations, can be set to NULL

join_names

whether to join concept names onto ids, default TRUE, FALSE used by omop_relations() to speed up

messages

whether to print info messages, default=TRUE

Value

a dataframe of concepts and attributes

Examples

omop_relations1step("Non-invasive blood pressure")
#>    querying concept relations of: Non-invasive blood pressure - may take a few seconds
#> returning 8 concepts
#> # A tibble: 8 × 13
#>   concept_id_1 concept_name_1        concept_id_2 concept_name_2 relationship_id
#>          <int> <chr>                        <int> <chr>          <chr>          
#> 1     36716965 Non-invasive blood p…     40642538 Primitive      Has status     
#> 2     36716965 Non-invasive blood p…     40642539 SNOMED CT core Has Module     
#> 3     36716965 Non-invasive blood p…      4326744 Blood pressure Is a           
#> 4     36716965 Non-invasive blood p…      4354254 Non-invasive … Subsumes       
#> 5     36716965 Non-invasive blood p…     36716283 Pressure       Has property   
#> 6     36716965 Non-invasive blood p…      4014241 Structure of … Inheres in     
#> 7     36716965 Non-invasive blood p…     36717771 Cardiac proce… Characterizes  
#> 8     36716965 Non-invasive blood p…      4149267 Quantitative   Has scale type 
#> # ℹ 8 more variables: valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>, domain_id <chr>, vocabulary_id <chr>,
#> #   concept_class_id <chr>, standard_concept <chr>, concept_code <chr>