extract omop concept relations of a vector of concept_ids - immediate relations with indication of relationship

extract omop concept relations of a vector of concept_ids - immediate relations with indication of relationship

Usage

omop_relations_multiple(
  mc_ids,
  c_ids = NULL,
  d_ids = NULL,
  v_ids = NULL,
  cc_ids = NULL,
  standard = NULL,
  r_ids = NULL,
  itself = FALSE,
  messages = TRUE,
  nsteps = 1
)

Arguments

mc_ids

a vector of multiple concept_id's to get relations of

c_ids

one or more concept_id to filter by, default NULL for all

d_ids

one or more domain_id to filter by, default NULL for all

v_ids

one or more vocabulary_id to filter by, default NULL for all

cc_ids

one or more concept_class_id to filter by, default NULL for all

standard

one or more standard_concept to filter by, default NULL for all, S,C

r_ids

one or more relationship_id to filter by, default NULL for all, e.g c('Is a','Subsumes')

itself

whether to include relations to concept itself, default=FALSE

messages

whether to print info messages, default=TRUE

nsteps

number of recursions to search

Value

a dataframe of concepts and attributes

Examples

orm <- omop_relations_multiple(c(3571338L,3655355L), r_ids=c('Is a','Subsumes'), nsteps=1)
#> multiple-y querying concept relations of: 2 concepts - may take more than a few seconds
#> recursively querying relations of: Problem behaviour 1/2
#> step 1 of 1
#>    querying concept relations of: Problem behaviour - may take a few seconds
#> returning 0 concepts
#> recursively querying relations of: Erectile dysfunction 2/2
#> step 1 of 1
#>    querying concept relations of: Erectile dysfunction - may take a few seconds
#> returning 16 concepts
#omop_relations_multiple(c(3571338L,3655355L), r_ids=c('Is a','Subsumes'), nsteps=2)