save omop vocabulary parquet file from provided location to local package cache

default using UCLH github repo, option to use local files

Usage

omop_vocab_table_save(
  tablename = "concept",
  from = NULL,
  to = tools::R_user_dir("omopcept", which = "cache"),
  download_max_secs = 720
)

Arguments

tablename

which omop table to download, defaults to 'concept'

from

url of file source location, defaults to UCLH filestore

to

path to save file locally, defaults to package cache which is where omopcept functions look

download_max_secs

max download secounds, default 720 = 12 mins

Value

nothing

See also