count freq of values in an omop field EXPERIMENTAL may change

count freq of values in an omop field EXPERIMENTAL may change

super short name func to count freq of values in an omop field

Usage

omopfreqrelationship(df1, sort = TRUE)

ofreqr(df1, sort = TRUE)

Arguments

df1

an omop table

sort

whether to sort freqs, default TRUE

Examples

omop_relations("Non-invasive blood pressure") |> omopfreqrelationship()
#> recursively querying concept relations of: Non-invasive blood pressure - may take more than a few seconds
#> step 1 of 1
#>    querying concept relations of: Non-invasive blood pressure - may take a few seconds
#> returning 8 concepts
#> # A tibble: 9 × 2
#>   relationship_id     n
#>   <chr>           <int>
#> 1 Characterizes       1
#> 2 Has Module          1
#> 3 Has property        1
#> 4 Has scale type      1
#> 5 Has status          1
#> 6 Inheres in          1
#> 7 Is                  1
#> 8 Is a                1
#> 9 Subsumes            1